feat: change to float

+++ include/models.h
@@ -7,8 +7,8 @@ class PayloadData
 {
 private:
     uint32_t logId;
-    uint32_t value;
-    uint32_t kalmanValue;
+    float value;
+    float kalmanValue;
 
 public:
     PayloadData() {};
@@ -18,11 +18,11 @@ public:
     {
         this->logId = logId;
     }
-    void setValue(const uint32_t &value)
+    void setValue(const float &value)
     {
         this->value = value;
     }
-    void setValueKalman(const uint32_t &value)
+    void setValueKalman(const float &value)
     {
         this->kalmanValue = value;
     }
